simple-tile: reorganize ipc handlers to a separate file
