+++ plugins/tile/tile-plugin.cpp
@@ -8,7 +8,6 @@
 #include <wayfire/signal-definitions.hpp>
 #include <wayfire/toplevel-view.hpp>
 
-#include "ipc-helpers.hpp"
 #include "plugins/ipc/ipc-method-repository.hpp"
 #include "tree-controller.hpp"
 #include "tree.hpp"
@@ -26,6 +25,7 @@
 #include "wayfire/view.hpp"
 
 #include "tile-wset.hpp"
+#include "tile-ipc.hpp"
 
 static bool can_tile_view(wayfire_toplevel_view view)
 {
@@ -465,7 +465,6 @@ class tile_plugin_t : public wf::plugin_interface_t, wf::per_output_tracker_mixi
         }
     };
 
-
     wf::signal::connection_t<view_moved_to_wset_signal> on_view_moved_to_wset =
         [=] (view_moved_to_wset_signal *ev)
     {
@@ -488,182 +487,13 @@ class tile_plugin_t : public wf::plugin_interface_t, wf::per_output_tracker_mixi
 
     ipc::method_callback ipc_get_layout = [=] (const nlohmann::json& params)
     {
-        WFJSON_EXPECT_FIELD(params, "wset-index", number_unsigned);
-        WFJSON_EXPECT_FIELD(params, "workspace", object);
-        WFJSON_EXPECT_FIELD(params["workspace"], "x", number_unsigned);
-        WFJSON_EXPECT_FIELD(params["workspace"], "y", number_unsigned);
-
-        int x   = params["workspace"]["x"].get<int>();
-        int y   = params["workspace"]["y"].get<int>();
-        auto ws = ipc::find_workspace_set_by_index(params["wset-index"].get<int>());
-        if (ws)
-        {
-            auto grid_size = ws->get_workspace_grid_size();
-            if ((x >= grid_size.width) || (y >= grid_size.height))
-            {
-                return wf::ipc::json_error("invalid workspace coordinates");
-            }
-
-            auto response = wf::ipc::json_ok();
-
-            auto cur_ws     = ws->get_current_workspace();
-            auto resolution = ws->get_last_output_geometry().value_or(tile::default_output_resolution);
-            wf::point_t offset = {cur_ws.x * resolution.width, cur_ws.y * resolution.height};
-
-            response["layout"] =
-                tree_to_json(tile_workspace_set_data_t::get(ws->shared_from_this()).roots[x][y], offset);
-            return response;
-        }
-
-        return wf::ipc::json_error("wset-index not found");
+        return tile::handle_ipc_get_layout(params);
     };
 
     ipc::method_callback ipc_set_layout = [=] (nlohmann::json params) -> nlohmann::json
     {
-        WFJSON_EXPECT_FIELD(params, "wset-index", number_unsigned);
-        WFJSON_EXPECT_FIELD(params, "workspace", object);
-        WFJSON_EXPECT_FIELD(params["workspace"], "x", number_unsigned);
-        WFJSON_EXPECT_FIELD(params["workspace"], "y", number_unsigned);
-        WFJSON_EXPECT_FIELD(params, "layout", object);
-        int x = params["workspace"]["x"].get<int>();
-        int y = params["workspace"]["y"].get<int>();
-
-        auto ws = ipc::find_workspace_set_by_index(params["wset-index"].get<int>());
-        if (!ws)
-        {
-            return wf::ipc::json_error("wset-index not found");
-        }
-
-        auto grid_size = ws->get_workspace_grid_size();
-        if ((x >= grid_size.width) || (y >= grid_size.height))
-        {
-            return wf::ipc::json_error("invalid workspace coordinates");
-        }
-
-        auto& tile_ws = tile_workspace_set_data_t::get(ws->shared_from_this());
-
-        tile::json_builder_data_t data;
-        data.gaps = tile_ws.get_gaps();
-        auto workarea = tile_ws.roots[x][y]->geometry;
-        if (auto err = tile::verify_json_tree(params["layout"], data, wf::dimensions(workarea)))
-        {
-            return wf::ipc::json_error(*err);
-        }
-
-        // Step 1: detach any views which are currently present in the layout, but should no longer be
-        // in the layout
-        std::vector<nonstd::observer_ptr<tile::view_node_t>> views_to_remove;
-        tile::for_each_view(tile_ws.roots[x][y], [&] (wayfire_toplevel_view view)
-        {
-            if (!data.touched_views.count(view))
-            {
-                views_to_remove.push_back(tile::view_node_t::get_node(view));
-            }
-        });
-
-        tile_ws.detach_views(views_to_remove);
-
-        {
-            autocommit_transaction_t tx;
-            data.touched_wsets.erase(nullptr);
-
-            // Step 2: temporarily detach some of the nodes
-            for (auto& touched_view : data.touched_views)
-            {
-                auto tile = wf::tile::view_node_t::get_node(touched_view);
-                if (tile)
-                {
-                    tile->parent->remove_child(tile, tx.tx);
-                }
-
-                if (touched_view->get_wset().get() != ws)
-                {
-                    auto old_wset = touched_view->get_wset();
-                    wf::emit_view_pre_moved_to_wset_pre(touched_view,
-                        touched_view->get_wset(), ws->shared_from_this());
-
-                    if (old_wset)
-                    {
-                        old_wset->remove_view(touched_view);
-                    }
-
-                    ws->add_view(touched_view);
-                    wf::emit_view_moved_to_wset(touched_view, old_wset, ws->shared_from_this());
-                }
-            }
-
-            // Step 3: set up the new layout
-            tile_ws.roots[x][y] = build_tree_from_json(params["layout"], &tile_ws, {x, y});
-            tile::flatten_tree(tile_ws.roots[x][y]);
-            tile_ws.roots[x][y]->set_gaps(tile_ws.get_gaps());
-            tile_ws.roots[x][y]->set_geometry(workarea, tx.tx);
-        }
-
-        data.touched_wsets.insert(ws);
-
-        // Step 4: flatten roots, set gaps, trigger resize everywhere
-        for (auto& touched_ws : data.touched_wsets)
-        {
-            auto& tws = tile_workspace_set_data_t::get(touched_ws->shared_from_this());
-            tws.flatten_roots();
-            // will also trigger resize everywhere
-            tws.update_gaps();
-        }
-
-        return wf::ipc::json_ok();
+        return tile::handle_ipc_set_layout(params);
     };
-
-    /**
-     * Build a tiling tree from a json description.
-     *
-     * Note that the tree description first has to be verified and pre-processed by verify_json_tree().
-     */
-    std::unique_ptr<tile::tree_node_t> build_tree_from_json(const nlohmann::json& json,
-        tile_workspace_set_data_t *wdata, wf::point_t vp)
-    {
-        auto root = build_tree_from_json_rec(json, wdata, vp);
-        if (root->as_view_node())
-        {
-            // Handle cases with a single view.
-            auto split_root = std::make_unique<tile::split_node_t>(tile_workspace_set_data_t::default_split);
-            split_root->children.push_back(std::move(root));
-            return split_root;
-        }
-
-        return root;
-    }
-
-    std::unique_ptr<tile::tree_node_t> build_tree_from_json_rec(const nlohmann::json& json,
-        tile_workspace_set_data_t *wdata, wf::point_t vp)
-    {
-        std::unique_ptr<tile::tree_node_t> root;
-
-        if (json.count("view-id"))
-        {
-            auto view = toplevel_cast(wf::ipc::find_view_by_id(json["view-id"]));
-            root = wdata->setup_view_tiling(view, vp);
-        } else
-        {
-            const bool is_horiz_split = json.count("horizontal-split");
-            auto& children_list = is_horiz_split ? json["horizontal-split"] : json["vertical-split"];
-            auto split_parent   = std::make_unique<tile::split_node_t>(
-                is_horiz_split ? tile::SPLIT_HORIZONTAL : tile::SPLIT_VERTICAL);
-
-            for (auto& child : children_list)
-            {
-                split_parent->children.push_back(build_tree_from_json_rec(child, wdata, vp));
-                split_parent->children.back()->parent = {split_parent.get()};
-            }
-
-            root = std::move(split_parent);
-        }
-
-        root->geometry.x     = 0;
-        root->geometry.y     = 0;
-        root->geometry.width = json["width"];
-        root->geometry.height = json["height"];
-        return root;
-    }
 };
 }
