+++ plugins/tile/tree-controller.cpp
@@ -1,5 +1,4 @@
 #include "tree-controller.hpp"
-#include "plugins/ipc/ipc-helpers.hpp"
 
 #include <set>
 #include <wayfire/nonstd/tracking-allocator.hpp>
@@ -568,138 +567,6 @@ void resize_view_controller_t::input_motion(wf::point_t input)
 
     wf::get_core().tx_manager->schedule_transaction(std::move(tx));
     this->last_point = input;
-}
-
-nlohmann::json tree_to_json(const std::unique_ptr<tree_node_t>& root, const wf::point_t& offset,
-    double rel_size)
-{
-    nlohmann::json js;
-    js["percent"]  = rel_size;
-    js["geometry"] = wf::ipc::geometry_to_json(root->geometry - offset);
-    if (auto view = root->as_view_node())
-    {
-        js["view-id"] = view->view->get_id();
-        return js;
-    }
-
-    auto split = root->as_split_node();
-    wf::dassert(split != nullptr, "Expected to be split node");
-
-    nlohmann::json children = nlohmann::json::array();
-    if (split->get_split_direction() == SPLIT_HORIZONTAL)
-    {
-        for (auto& child : split->children)
-        {
-            children.push_back(
-                tree_to_json(child, offset, 1.0 * child->geometry.height / split->geometry.height));
-        }
-
-        js["horizontal-split"] = std::move(children);
-    } else
-    {
-        for (auto& child : split->children)
-        {
-            children.push_back(tree_to_json(
-                child, offset, 1.0 * child->geometry.width / split->geometry.width));
-        }
-
-        js["vertical-split"] = std::move(children);
-    }
-
-    return js;
-}
-
-std::optional<std::string> verify_json_tree(nlohmann::json& json, json_builder_data_t& data,
-    const wf::dimensions_t& available_geometry)
-{
-    if (!json.is_object())
-    {
-        return "JSON Tree structure is wrong!";
-    }
-
-    if ((available_geometry.width <= data.gaps.left + data.gaps.right) ||
-        (available_geometry.height <= data.gaps.top + data.gaps.bottom))
-    {
-        return "Geometry becomes too small for some nodes!";
-    }
-
-    json["width"]  = available_geometry.width;
-    json["height"] = available_geometry.height;
-    if (json.count("view-id"))
-    {
-        if (!json["view-id"].is_number_unsigned())
-        {
-            return "view-id should be unsigned integer!";
-        }
-
-        auto view = toplevel_cast(wf::ipc::find_view_by_id(json["view-id"]));
-        if (!view)
-        {
-            return "No view found with id " + std::to_string((uint32_t)json["view-id"]);
-        }
-
-        if (data.touched_views.count(view))
-        {
-            return "View tiled twice!";
-        }
-
-        data.touched_views.insert(view);
-        data.touched_wsets.insert(view->get_wset().get());
-        return {};
-    }
-
-    const bool is_horiz_split = json.count("horizontal-split") && json["horizontal-split"].is_array();
-    const bool is_vert_split  = json.count("vertical-split") && json["vertical-split"].is_array();
-    if (!is_horiz_split && !is_vert_split)
-    {
-        return "Node is neither a view, nor a split node!";
-    }
-
-    int32_t split_axis = is_horiz_split ? available_geometry.height : available_geometry.width;
-    double weight_sum  = 0;
-
-    auto& children_list = is_horiz_split ? json["horizontal-split"] : json["vertical-split"];
-
-    for (auto& child : children_list)
-    {
-        if (!child.count("weight"))
-        {
-            return "Expected 'weight' field for each child node!";
-        }
-
-        if (!child["weight"].is_number())
-        {
-            return "Expected 'weight' field to be a number!";
-        }
-
-        weight_sum += float(child["weight"]);
-    }
-
-    int32_t size_sum = 0;
-    for (auto& child : children_list)
-    {
-        int32_t size = float(child["weight"]) / weight_sum * split_axis;
-        size_sum += size;
-        if (&child == &children_list.back())
-        {
-            // This is needed because of rounding errors, we always round down, but in the end, we need to
-            // make sure that the nodes cover the whole screen.
-            size += split_axis - size_sum;
-        }
-
-        const wf::dimensions_t available_for_child = is_horiz_split ?
-            wf::dimensions_t{available_geometry.width, size} :
-        wf::dimensions_t{size, available_geometry.height};
-
-        auto error = verify_json_tree(child, data, available_for_child);
-        if (error.has_value())
-        {
-            return error;
-        }
-    }
-
-    // All was OK
-    return {};
 }
 } // namespace tile
 }
